Fix problem when ruby exe path has spaces
On Windows, it's possible that the RbConfig path to the ruby executable
might contain spaces (e.g. "C:/Program Files/foo/ruby.exe"). If that
happens, systemu's previous method of checking to see if that was a
valid ruby executable was unreliable. This commit allows full paths to
executables to have spaces by switching to individual argument
invocation of system(), instead of relying on system() to correctly
parse a single string into executable + arguments.